+++ tests/test_coding/test_structures.py
@@ -523,8 +523,9 @@ class TestStructures(unittest.TestCase):
             def __init__(self, v):
                 self.v = v
 
-        self.assertTrue(repr(Test2(2)).endswith(
-            'test_coding.test_structures.TestMisc.test_reprable_mixin.<locals>.Test2(2)'))
+        self.assertEqual(repr(Test2(2)),
+                         'tests.test_coding.test_structures.TestStructures.' \
+                         'test_reprable_mixin.<locals>.Test2(2)')
 
     def test_proxy(self):
         a = Proxy(5, wrap_operations=True)