+++ satella/coding/resources/__init__.py
@@ -1,0 +1,1 @@
+from .cp_manager import CPManager