+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # v2.18.11
+
+* unit tests migrated to CircleCI