+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # v2.18.11
 
 * unit tests migrated to CircleCI
+* added __len__ to FutureCollection