+++ CHANGELOG.md
@@ -2,3 +2,5 @@
 
 * unit tests migrated to CircleCI
 * added __len__ to FutureCollection
+* fixed a bug in DictionaryEQAble
+* fixed a bug in ListDeleter