+++ CHANGELOG.md
@@ -1,7 +1,8 @@
-# v2.18.11
+# v2.19.0
 
 * unit tests migrated to CircleCI
 * added __len__ to FutureCollection
 * fixed a bug in DictionaryEQAble
 * fixed a bug in ListDeleter
 * minor breaking change: changed semantics of ListDeleter
+* added `CPManager`